\name{vsn}
\docType{class}
\alias{class:vsn}
\alias{vsn-class}


\title{
  Class to contain result of a vsn fit
}

\description{
  Class to contain result of a vsn fit
}

\section{Creating Objects}{
  \code{new("vsn")}
  \code{vsn2("ExpressionSet")},
}

\section{Slots}{
  TO DO
}
\section{Methods}{
  TO DO
}

\author{Wolfgang Huber \url{http://www.ebi.ac.uk/huber}}

\seealso{
  \code{\link{vsn2}}
}

\examples{
}

\keyword{classes}
