\name{vsnInput}
\docType{class}
\alias{class:vsnInput}
\alias{vsnInput-class}


\title{
  Class to contain input data and parameters for the model
}

\description{
  Class to contain input data and parameters for the model
}

\section{Creating Objects}{
  \code{new("vsnInput")}
}

\section{Slots}{
  TO DO
}
\section{Methods}{
  TO DO
}

\author{Wolfgang Huber \url{http://www.ebi.ac.uk/huber}}

\seealso{
  \code{\link{vsn2}}
}

\examples{
}

\keyword{classes}
