\name{sqr}
\alias{sqr}
\title{Square of a numeric expression}
\description{Square of a numeric expression}
\usage{
sqr(x)
}
\arguments{\item{x}{Numeric.}}
\value{x*x}
\details{Faster than \code{x^2}.
}
\author{Wolfgang Huber \url{http://www.dkfz.de/abt0840/whuber}}
\seealso{\code{\link{sqrt}}}

\examples{
  sqr(2)
}

\keyword{math}
